+++ +sbfsem/+core/Ultrastructure.m
@@ -1,16 +1,33 @@
 classdef (Abstract) Ultrastructure < handle
+% ULTRASTRUCTURE
+%
+% Description:
+%   Abstract class for ultrastructure annotations (mitochondria, axon, etc)
+%
+% Properties:
+%   source          Volume name
+%   baseURL         Service root for OData queries
+% Protected properties:
+%   queryDate       Time of last update from OData
+%   TYPEID          Viking StructureType ID - must be defined by subclasses
+%
+% History:
+%   5Jan - SSP - created
+% -------------------------------------------------------------------------
     
     properties (SetAccess = private, GetAccess = public)
         source
+    end
+
+    properties (SetAccess = protected, GetAccess = public)
         queryDate
     end
     
-    properties (Access = private)
+    properties (Access = protected)
         baseURL
     end
     
     properties (Constant = true, Abstract = true)
-        % All subclasses must define a Viking TypeID
         TYPEID
     end